Fix shell injection/XSS/NameError issues + v1.9.4

+++ setup.sh
@@ -157,7 +157,7 @@ fi
 # --- Deploy app from GitHub ---
 log "Cloning serv-UI from GitHub..."
 TEMP_DIR=$(mktemp -d)
-trap "rm -rf $TEMP_DIR" EXIT
+trap 'rm -rf "${TEMP_DIR:?}"' EXIT
 
 if [[ -d "$APP_DIR/.git" ]]; then
   info "Existing installation found. Updating..."
